+++ packages/rest-api/src/tests/bridgeLimitsRoute.test.ts
@@ -1,14 +1,14 @@
 import request from 'supertest'
 import express from 'express'
 
-import getBridgeLimitsRoute from '../routes/bridgeLimitsRoute'
+import bridgeLimitsRoute from '../routes/bridgeLimitsRoute'
 
 const app = express()
-app.use('/getBridgeLimits', getBridgeLimitsRoute)
+app.use('/bridgeLimits', bridgeLimitsRoute)
 
 describe('Get Bridge Limits Route', () => {
   it('should return min/max origin amounts for valid input', async () => {
-    const response = await request(app).get('/getBridgeLimits').query({
+    const response = await request(app).get('/bridgeLimits').query({
       fromChain: 1,
       fromToken: '0xA0b86991c6218b36c1d19D4a2e9Eb0cE3606eB48',
       toChain: 10,
@@ -21,7 +21,7 @@ describe('Get Bridge Limits Route', () => {
   }, 10_000)
 
   it('should return 400 for unsupported fromChain', async () => {
-    const response = await request(app).get('/getBridgeLimits').query({
+    const response = await request(app).get('/bridgeLimits').query({
       fromChain: '999',
       toChain: '137',
       fromToken: '0x176211869cA2b568f2A7D4EE941E073a821EE1ff',
@@ -35,7 +35,7 @@ describe('Get Bridge Limits Route', () => {
   }, 10_000)
 
   it('should return 400 for unsupported ', async () => {
-    const response = await request(app).get('/getBridgeLimits').query({
+    const response = await request(app).get('/bridgeLimits').query({
       fromChain: '999',
       toChain: '137',
       fromToken: '0x176211869cA2b568f2A7D4EE941E073a821EE1ff',
@@ -49,7 +49,7 @@ describe('Get Bridge Limits Route', () => {
   }, 10_000)
 
   it('should return 400 for unsupported toChain', async () => {
-    const response = await request(app).get('/getBridgeLimits').query({
+    const response = await request(app).get('/bridgeLimits').query({
       fromChain: '137',
       toChain: '999',
       fromToken: '0x3c499c542cEF5E3811e1192ce70d8cC03d5c3359',
@@ -60,7 +60,7 @@ describe('Get Bridge Limits Route', () => {
   }, 10_000)
 
   it('should return 400 for missing fromToken', async () => {
-    const response = await request(app).get('/getBridgeLimits').query({
+    const response = await request(app).get('/bridgeLimits').query({
       fromChain: '1',
       toChain: '137',
       toToken: '0x176211869cA2b568f2A7D4EE941E073a821EE1ff',
@@ -70,7 +70,7 @@ describe('Get Bridge Limits Route', () => {
   }, 10_000)
 
   it('should return 400 for missing toToken', async () => {
-    const response = await request(app).get('/getBridgeLimits').query({
+    const response = await request(app).get('/bridgeLimits').query({
       fromChain: '1',
       toChain: '137',
       fromToken: '0xa0b86991c6218b36c1d19d4a2e9eb0ce3606eb48',
